+++ src/Products/MeetingCommunes/config.py
@@ -7,6 +7,7 @@
 
 from Products.CMFCore.permissions import setDefaultRoles
 from Products.PloneMeeting import config
+from Products.PloneMeeting.profiles import CategoryDescriptor
 
 
 __author__ = """Gauthier Bastien <g.bastien@imio.be>, Stephan Geulette <s.geulette@imio.be>"""
@@ -66,3 +67,35 @@ SAMPLE_TEXT = u"<p><strong>Lorem ipsum dolor sit amet</strong>, consectetur adip
 MC_ITEM_TRANSITION_WHEN_RETURNED_FROM_PROPOSING_GROUP_AFTER_CORRECTION = 'accept_but_modify'
 config.ITEM_TRANSITION_WHEN_RETURNED_FROM_PROPOSING_GROUP_AFTER_CORRECTION = \
     MC_ITEM_TRANSITION_WHEN_RETURNED_FROM_PROPOSING_GROUP_AFTER_CORRECTION
+
+PORTAL_CATEGORIES = [
+    CategoryDescriptor("administration", "Administration générale"),
+    CategoryDescriptor("immo", "Affaires immobilières"),
+    CategoryDescriptor("espaces-publics", "Aménagement des espaces publics"),
+    CategoryDescriptor("batiments-communaux", "Bâtiments communaux"),
+    CategoryDescriptor("animaux", "Bien-être animal"),
+    CategoryDescriptor("communication", "Communication & Relations extérieures"),
+    CategoryDescriptor("cultes", "Cultes"),
+    CategoryDescriptor("culture", "Culture & Folklore"),
+    CategoryDescriptor("economie", "Développement économique & commercial"),
+    CategoryDescriptor("enseignement", "Enseignement"),
+    CategoryDescriptor("population", "État civil & Population"),
+    CategoryDescriptor("finances", "Finances"),
+    CategoryDescriptor("informatique", "Informatique"),
+    CategoryDescriptor("interculturalite", "Interculturalité & Égalité"),
+    CategoryDescriptor("jeunesse", "Jeunesse"),
+    CategoryDescriptor("logement", "Logement & Énergie"),
+    CategoryDescriptor("mobilite", "Mobilité"),
+    CategoryDescriptor("quartier", "Participation relation avec les quartiers"),
+    CategoryDescriptor("patrimoine", "Patrimoine"),
+    CategoryDescriptor("enfance", "Petite enfance"),
+    CategoryDescriptor("politique", "Politique générale"),
+    CategoryDescriptor("environnement", "Propreté & Environnement"),
+    CategoryDescriptor("sante", "Santé"),
+    CategoryDescriptor("securite", "Sécurité & Prévention"),
+    CategoryDescriptor("social", "Services sociaux"),
+    CategoryDescriptor("sport", "Sport"),
+    CategoryDescriptor("tourisme", "Tourisme"),
+    CategoryDescriptor("urbanisme", "Urbanisme & Aménagement du territoire"),
+    CategoryDescriptor("police", "Zone de police"),
+]